Accept integers in programmatically created filter/predicate arguments. Issue no : 296
Modified eskip to parse numbers as int64 too.
Some filters and predicates relying this accept only float64 args in number positions.
This is very inconvenient when creating filters programmatically.
Where possible fixed these filters to accepts integers (int64), too.

+++ innkeeper/client_test.go
@@ -569,7 +569,7 @@ func TestUsesPreAndPostRouteFilters(t *testing.T) {
 		if r.Filters[1].Name != "filter2" ||
 			len(r.Filters[1].Args) != 2 ||
 			r.Filters[1].Args[0] != "key" ||
-			r.Filters[1].Args[1] != float64(42) {
+			r.Filters[1].Args[1] != int64(42) {
 			t.Error("failed to parse filters 3")
 		}
